requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.24.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
filterpy==1.4.5
python-dotenv==1.0.0
supabase==2.0.2
streamlit==1.28.2
python-multipart==0.0.6
PyYAML==6.0.1
